copy Question Item
