+++ knowledge/src/knowledgeTypes.ts
@@ -102,7 +102,6 @@ export type KnowledgeNode = Omit<
   tags?: LinkedKnowledgeNode[];
   createdAt?: string;
   choices?: KnowledgeChoice[];
-  constent?: string;
   references?: LinkedKnowledgeNode[];
   contributors?: KnowledgeNodeContributor[];
   institutions?: KnowledgeNodeInstitution[];
